fix: avoid unknown destroy of texture the intermediate node
