+++ Editor/NDMF/TexTransPhaseNode.cs.meta
@@ -1,0 +1,11 @@
+fileFormatVersion: 2
+guid: d5913b8819331f140a5b7e3aeb9de195
+MonoImporter:
+  externalObjects: {}
+  serializedVersion: 2
+  defaultReferences: []
+  executionOrder: 0
+  icon: {instanceID: 0}
+  userData: 
+  assetBundleName: 
+  assetBundleVariant:
